mrpt-nav done!

+++ libs/nav/src/planners/PlannerRRT_common.cpp
@@ -113,7 +113,7 @@ void PlannerTPS_VirtualBase::internal_loadConfig_PTG(
 			ASSERT_(mShape.rows() == 2);
 			ASSERT_(mShape.cols() >= 3);
 
-			for (size_t i = 0; i < mShape.cols(); i++)
+			for (int i = 0; i < mShape.cols(); i++)
 				params.robot_shape.push_back(
 					TPoint2D(mShape(0, i), mShape(1, i)));
 		}
